+++ tilitili/app/src/main/java/com/example/tilitili/ui/ActivityFragment.java
@@ -84,7 +84,7 @@ public class ActivityFragment extends BaseFragment implements Pager.OnPageListen
                                 item.getString("cover"),
                                 item.getString("introduction"),
                                 item.getString("resource"),
-                                item.getInt("submissionTime"),
+                                item.getLong("submissionTime"),
                                 item.getInt("watchTimes"),
                                 item.getInt("likesCount"),
                                 item.getInt("isLike"),
